Set default-directory and file-name-handler-alist
This ensures that our behaviour is more deterministic, and prevents
loading URLs.

Functions like f-depth depend on the value of default-directory. If we
don't set this value, we use whatever value default-directory had in
the buffer that we ran 'M-x suggest' in. This hurts reproducibility.

Functions like expand-file-name depend on the value of
file-name-handler-alist. By default, this converts URLs to tramp
paths. This is problematic when we call f-entries, as we then try to
load URLs! suggest.el should not produce side-effects.

Fixes #32.