Fix erroneous candidate "quote" to cl-case
Passing 'foo as a candidate keylist to cl-case equates to passing
(quote foo), which actually means "match the symbol quote or foo". This
triggers a new-ish warning in Emacs 30. Although the form in question is
matching on the value of suggest--classify-output, which does not ever
return the symbol "quote", this is still worth fixing.

Fixes #51.